sg: move RoundTripper to subpackage
