+++ monitoring/monitoring/internal/headertransport/headertransport.go
@@ -1,4 +1,4 @@
-package monitoring
+package headertransport
 
 import "net/http"
